moved download attachment to libs vault components
