+++ libs/vault/src/components/download-attachment/download-attachment.component.html
